requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.2
django-crispy-forms-registration==0.1.3
django-phonenumber-field==5.0.0
django-registration==2.2
easy-thumbnails==2.7.1
gunicorn==20.1.0
numpy==1.20.1
phonenumbers==8.12.20
Pillow==8.1.2
pycifstar==0.2.8
pytz==2021.1
scipy==1.6.1
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2==2.8.6
